test: correct created tests
