+++ src/components/_view/MobileView/MobileView.test.tsx
@@ -1,30 +1,27 @@
 import '@testing-library/jest-dom';
 
 import { render } from '@testing-library/react';
-import { FC } from 'react';
 
 import { MobileView, MobileViewInfoText } from '_view/MobileView/MobileView';
-
-import { MIN_SYSTEM_RESOLUTION } from '../../constants';
+import { MIN_SYSTEM_RESOLUTION } from 'src/constants';
 
 type Props = {
   screenWidth: number;
 };
 
 describe('MobileView Component', () => {
-  test('Component should render a message', () => {
+  it('should render a not supported resolution message', () => {
     const { getByText } = render(<MobileView />);
     expect(getByText(MobileViewInfoText)).toBeInTheDocument();
   });
 
-  test(`Component should render only if browser resolution is less than or equal to ${MIN_SYSTEM_RESOLUTION}`, () => {
+  it(`should render only if browser resolution is less than or equal to ${MIN_SYSTEM_RESOLUTION}`, () => {
     const testId = 'mobile-view';
-
     const screenWidthEqualToMinResolution = MIN_SYSTEM_RESOLUTION;
     const screenWidthMoreThanMinResolution = MIN_SYSTEM_RESOLUTION + 500;
     const screenWidthLessThanMinResolution = MIN_SYSTEM_RESOLUTION - 500;
 
-    const RenderMobileView: FC<Props> = ({ screenWidth }) =>
+    const RenderMobileView = ({ screenWidth }: Props) =>
       screenWidth <= MIN_SYSTEM_RESOLUTION ? <MobileView /> : null;
 
     const { rerender, queryByTestId } = render(
