+++ src/hooks/useMountTransition/useMountTransition.test.ts
@@ -2,18 +2,14 @@ import { renderHook } from '@testing-library/react';
 
 import { useMountTransition } from 'hooks/useMountTransition';
 
-describe('useMountTransition hook tests', () => {
-  test('hook should return true as transition is handling', () => {
-    console.log(
-      // hookResult,
-      renderHook(() => useMountTransition(true, 0)),
-    );
+describe('useMountTransition hook', () => {
+  it('should return true if transition is handling', () => {
     const hookResult = renderHook(() => useMountTransition(true, 0));
 
     expect(hookResult.result.current).toBeTruthy();
   });
 
-  test('hook should return false as transition is not longer handling', () => {
+  test('hook should return false if transition is not longer handling', () => {
     const hookResult = renderHook(() => useMountTransition(false, 0));
     expect(hookResult.result.current).not.toBeTruthy();
   });
