+++ src/hooks/useScreenWidth/useScreenWidth.test.tsx
@@ -13,7 +13,7 @@ const EffectComponent = () => {
   return <span>{width}</span>;
 };
 
-test('useScreenWidth listen to window resize and set screen width to current screen size', () => {
+test('if useScreenWidth listen to window resize and setting screen width to current screen size', () => {
   const { container, rerender } = render(<EffectComponent />);
   const spanElement = container.firstChild;
 
@@ -21,15 +21,7 @@ test('useScreenWidth listen to window resize and set screen width to current scr
   rerender(<EffectComponent />);
   expect(spanElement?.textContent).toBe('320');
 
-  act(() => fireResizeEvent(800));
-  rerender(<EffectComponent />);
-  expect(spanElement?.textContent).toBe('800');
-
   act(() => fireResizeEvent(1200));
   rerender(<EffectComponent />);
   expect(spanElement?.textContent).toBe('1200');
-
-  act(() => fireResizeEvent(1900));
-  rerender(<EffectComponent />);
-  expect(spanElement?.textContent).toBe('1900');
 });
